requirements.txt
beautifulsoup4==4.9.3
certifi==2021.5.30
chardet==4.0.0
icrawler==0.6.4
idna==3.2
lxml==4.6.3
Pillow==8.2.0
requests==2.25.1
six==1.16.0
soupsieve==2.2.1
urllib3==1.26.5
